+++ flexiapi/app/Account.php
@@ -68,12 +68,7 @@ class Account extends Authenticatable
     protected static function booted()
     {
         static::addGlobalScope('domain', function (Builder $builder) {
-            if (Auth::hasUser()) {
-                $user = Auth::user();
-                if (!$user->admin || !config('app.admins_manage_multi_domains')) {
-                    $builder->where('domain', config('app.sip_domain'));
-                }
-
+            if (Auth::hasUser() && Auth::user()->admin && config('app.admins_manage_multi_domains')) {
                 return;
             }