+++ flexiapi/app/Http/Controllers/Admin/AccountActivityController.php
@@ -20,18 +20,17 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
 
 use App\Account;
 
 class AccountActivityController extends Controller
 {
-    public function index(Request $request, Account $account)
+    public function index(int $accountId)
     {
         return view(
             'admin.account.activity.index',
             [
-                'account' => $account
+                'account' => Account::findOrFail($accountId)
             ]
         );
     }